[receiver/jmx] Move subpackages to internal
Resolves #4068
